--- wiremock/mappings/external_browser/successful_flow.json
@@ -1,78 +1,0 @@
-{
-  "mappings": [
-    {
-      "request": {
-        "urlPathPattern": "/session/authenticator-request",
-        "method": "POST"
-      },
-      "response": {
-        "status": 200,
-        "headers": {
-          "Content-Type": "application/json"
-        },
-        "jsonBody": {
-          "data": {
-            "ssoUrl": "http://127.0.0.1:9999/sso-stub",
-            "proofKey": "proof-key-123"
-          },
-          "code": null,
-          "message": null,
-          "success": true
-        }
-      }
-    },
-    {
-      "request": {
-        "urlPathPattern": "/session/v1/login-request.*",
-        "method": "POST",
-        "bodyPatterns": [
-          {
-            "equalToJson": {
-              "data": {
-                "AUTHENTICATOR": "EXTERNALBROWSER"
-              }
-            },
-            "ignoreExtraElements": true
-          }
-        ]
-      },
-      "response": {
-        "status": 200,
-        "headers": {
-          "Content-Type": "application/json"
-        },
-        "jsonBody": {
-          "data": {
-            "masterToken": "master token",
-            "token": "session token",
-            "validityInSeconds": 3600,
-            "masterValidityInSeconds": 14400,
-            "displayUserName": "EXTERNAL_BROWSER_USER",
-            "serverVersion": "8.48.0",
-            "firstLogin": false,
-            "remMeToken": null,
-            "remMeValidityInSeconds": 0,
-            "healthCheckInterval": 45,
-            "newClientForUpgrade": "3.12.3",
-            "sessionId": 1172562260499,
-            "parameters": [],
-            "sessionInfo": {
-              "databaseName": "TEST_DATABASE",
-              "schemaName": "TEST_JDBC",
-              "warehouseName": "TEST_XSMALL",
-              "roleName": "ANALYST"
-            },
-            "idToken": null,
-            "idTokenValidityInSeconds": 0,
-            "responseData": null,
-            "mfaToken": null,
-            "mfaTokenValidityInSeconds": 0
-          },
-          "code": null,
-          "message": null,
-          "success": true
-        }
-      }
-    }
-  ]
-}
